requirements.txt
Django==5.1.2
djangorestframework==3.15.2
django-cors-headers==4.5.0
django-environ==0.11.2
psycopg2==2.9.10
